# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Kampüsnet API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface UserProfile {
  id?: number;
  clerkId: string;
  email: string;
  firstName: string;
  lastName: string;
  university: string;
  department: string;
  year: number;
  /** @nullable */
  bio?: string | null;
  /** @nullable */
  avatarUrl?: string | null;
  postCount?: number;
  followerCount?: number;
  createdAt: string;
}

export interface UserProfileInput {
  /** @minLength 1 */
  firstName: string;
  /** @minLength 1 */
  lastName: string;
  /** @minLength 1 */
  university: string;
  /** @minLength 1 */
  department: string;
  /**
   * @minimum 1
   * @maximum 8
   */
  year: number;
  bio?: string;
  avatarUrl?: string;
}

export interface Post {
  id: number;
  content: string;
  /** @nullable */
  imageUrl?: string | null;
  authorId: string;
  authorName: string;
  authorUniversity: string;
  authorDepartment?: string;
  /** @nullable */
  authorAvatarUrl?: string | null;
  likeCount: number;
  commentCount: number;
  liked: boolean;
  createdAt: string;
}

export interface PostInput {
  /**
   * @minLength 1
   * @maxLength 2000
   */
  content: string;
  imageUrl?: string;
}

export interface PostsPage {
  posts: Post[];
  total: number;
}

export interface Comment {
  id: number;
  content: string;
  postId: number;
  authorId: string;
  authorName: string;
  authorUniversity: string;
  /** @nullable */
  authorAvatarUrl?: string | null;
  createdAt: string;
}

export interface CommentInput {
  /**
   * @minLength 1
   * @maxLength 1000
   */
  content: string;
}

export interface LikeResult {
  liked: boolean;
  likeCount: number;
}

export interface FeedStats {
  totalPosts: number;
  totalUsers: number;
  totalUniversities: number;
  postsToday: number;
}

export type GetPostsParams = {
  /**
   * Filter posts by university
   */
  filter?: GetPostsFilter;
  limit?: number;
  offset?: number;
};

export type GetPostsFilter =
  (typeof GetPostsFilter)[keyof typeof GetPostsFilter];

export const GetPostsFilter = {
  all: "all",
  my_university: "my_university",
} as const;