# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Kampüsnet API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: users
    description: User profile operations
  - name: posts
    description: Post operations
  - name: comments
    description: Comment operations
  - name: likes
    description: Like operations
  - name: feed
    description: Feed and stats operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /users/me:
    get:
      operationId: getMyProfile
      tags: [users]
      summary: Get current user profile
      responses:
        "200":
          description: Current user profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
        "401":
          description: Unauthorized
        "404":
          description: Profile not found
    put:
      operationId: upsertMyProfile
      tags: [users]
      summary: Create or update current user profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserProfileInput"
      responses:
        "200":
          description: Updated profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
        "400":
          description: Invalid input (e.g. non .edu.tr email)
        "401":
          description: Unauthorized

  /users/{userId}:
    get:
      operationId: getUserProfile
      tags: [users]
      summary: Get a user profile by ID
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: User profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
        "404":
          description: Not found

  /posts:
    get:
      operationId: getPosts
      tags: [posts]
      summary: Get posts feed
      parameters:
        - name: filter
          in: query
          schema:
            type: string
            enum: [all, my_university]
          description: Filter posts by university
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        "200":
          description: List of posts
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PostsPage"
        "401":
          description: Unauthorized
    post:
      operationId: createPost
      tags: [posts]
      summary: Create a new post
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PostInput"
      responses:
        "201":
          description: Created post
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Post"
        "401":
          description: Unauthorized

  /posts/{postId}:
    get:
      operationId: getPost
      tags: [posts]
      summary: Get a single post
      parameters:
        - name: postId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Post
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Post"
        "404":
          description: Not found
    delete:
      operationId: deletePost
      tags: [posts]
      summary: Delete a post (owner only)
      parameters:
        - name: postId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "404":
          description: Not found

  /posts/{postId}/like:
    post:
      operationId: likePost
      tags: [likes]
      summary: Like a post
      parameters:
        - name: postId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Like result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LikeResult"
        "401":
          description: Unauthorized
    delete:
      operationId: unlikePost
      tags: [likes]
      summary: Unlike a post
      parameters:
        - name: postId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Unlike result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LikeResult"
        "401":
          description: Unauthorized

  /posts/{postId}/comments:
    get:
      operationId: getComments
      tags: [comments]
      summary: Get comments for a post
      parameters:
        - name: postId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of comments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Comment"
        "404":
          description: Post not found
    post:
      operationId: createComment
      tags: [comments]
      summary: Add a comment to a post
      parameters:
        - name: postId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CommentInput"
      responses:
        "201":
          description: Created comment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Comment"
        "401":
          description: Unauthorized

  /comments/{commentId}:
    delete:
      operationId: deleteComment
      tags: [comments]
      summary: Delete a comment (owner only)
      parameters:
        - name: commentId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "404":
          description: Not found

  /feed/stats:
    get:
      operationId: getFeedStats
      tags: [feed]
      summary: Get feed statistics
      responses:
        "200":
          description: Stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FeedStats"
        "401":
          description: Unauthorized

  /feed/trending:
    get:
      operationId: getTrendingPosts
      tags: [feed]
      summary: Get trending posts (most liked in last 24h)
      responses:
        "200":
          description: Trending posts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Post"
        "401":
          description: Unauthorized

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    UserProfile:
      type: object
      required: [clerkId, email, firstName, lastName, university, department, year, createdAt]
      properties:
        id:
          type: integer
        clerkId:
          type: string
        email:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        university:
          type: string
        department:
          type: string
        year:
          type: integer
        bio:
          type: ["string", "null"]
        avatarUrl:
          type: ["string", "null"]
        postCount:
          type: integer
        followerCount:
          type: integer
        createdAt:
          type: string

    UserProfileInput:
      type: object
      required: [firstName, lastName, university, department, year]
      properties:
        firstName:
          type: string
          minLength: 1
        lastName:
          type: string
          minLength: 1
        university:
          type: string
          minLength: 1
        department:
          type: string
          minLength: 1
        year:
          type: integer
          minimum: 1
          maximum: 8
        bio:
          type: string
        avatarUrl:
          type: string

    Post:
      type: object
      required: [id, content, authorId, authorName, authorUniversity, likeCount, commentCount, liked, createdAt]
      properties:
        id:
          type: integer
        content:
          type: string
        imageUrl:
          type: ["string", "null"]
        authorId:
          type: string
        authorName:
          type: string
        authorUniversity:
          type: string
        authorDepartment:
          type: string
        authorAvatarUrl:
          type: ["string", "null"]
        likeCount:
          type: integer
        commentCount:
          type: integer
        liked:
          type: boolean
        createdAt:
          type: string

    PostInput:
      type: object
      required: [content]
      properties:
        content:
          type: string
          minLength: 1
          maxLength: 2000
        imageUrl:
          type: string

    PostsPage:
      type: object
      required: [posts, total]
      properties:
        posts:
          type: array
          items:
            $ref: "#/components/schemas/Post"
        total:
          type: integer

    Comment:
      type: object
      required: [id, content, postId, authorId, authorName, authorUniversity, createdAt]
      properties:
        id:
          type: integer
        content:
          type: string
        postId:
          type: integer
        authorId:
          type: string
        authorName:
          type: string
        authorUniversity:
          type: string
        authorAvatarUrl:
          type: ["string", "null"]
        createdAt:
          type: string

    CommentInput:
      type: object
      required: [content]
      properties:
        content:
          type: string
          minLength: 1
          maxLength: 1000

    LikeResult:
      type: object
      required: [liked, likeCount]
      properties:
        liked:
          type: boolean
        likeCount:
          type: integer

    FeedStats:
      type: object
      required: [totalPosts, totalUsers, totalUniversities, postsToday]
      properties:
        totalPosts:
          type: integer
        totalUsers:
          type: integer
        totalUniversities:
          type: integer
        postsToday:
          type: integer