# lib/api-zod/src/generated/types/comment.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Kampüsnet API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Comment {
  id: number;
  content: string;
  postId: number;
  authorId: string;
  authorName: string;
  authorUniversity: string;
  /** @nullable */
  authorAvatarUrl?: string | null;
  createdAt: string;
}