# lib/api-zod/src/generated/types/commentInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Kampüsnet API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CommentInput {
  /**
   * @minLength 1
   * @maxLength 1000
   */
  content: string;
}