# lib/api-zod/src/generated/types/feedStats.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Kampüsnet API specification
 * OpenAPI spec version: 0.1.0
 */

export interface FeedStats {
  totalPosts: number;
  totalUsers: number;
  totalUniversities: number;
  postsToday: number;
}