# lib/api-zod/src/generated/types/likeResult.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Kampüsnet API specification
 * OpenAPI spec version: 0.1.0
 */

export interface LikeResult {
  liked: boolean;
  likeCount: number;
}