# lib/api-zod/src/generated/types/postInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Kampüsnet API specification
 * OpenAPI spec version: 0.1.0
 */

export interface PostInput {
  /**
   * @minLength 1
   * @maxLength 2000
   */
  content: string;
  imageUrl?: string;
}