# lib/api-zod/src/generated/types/userProfile.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Kampüsnet API specification
 * OpenAPI spec version: 0.1.0
 */

export interface UserProfile {
  id?: number;
  clerkId: string;
  email: string;
  firstName: string;
  lastName: string;
  university: string;
  department: string;
  year: number;
  /** @nullable */
  bio?: string | null;
  /** @nullable */
  avatarUrl?: string | null;
  postCount?: number;
  followerCount?: number;
  createdAt: string;
}