# lib/api-zod/src/generated/types/userProfileInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Kampüsnet API specification
 * OpenAPI spec version: 0.1.0
 */

export interface UserProfileInput {
  /** @minLength 1 */
  firstName: string;
  /** @minLength 1 */
  lastName: string;
  /** @minLength 1 */
  university: string;
  /** @minLength 1 */
  department: string;
  /**
   * @minimum 1
   * @maximum 8
   */
  year: number;
  bio?: string;
  avatarUrl?: string;
}